+++ ios/Classes/AppsflyerSdkPlugin.m
@@ -3,7 +3,10 @@
 
 @implementation AppsflyerSdkPlugin {
     FlutterEventChannel *_eventChannel;
+    FlutterMethodChannel *_callbackChannel;
     AppsFlyerStreamHandler *_streamHandler;
+    // Callbacks
+    NSMutableArray* callbackById;
 }
 
 - (instancetype)initWithMessenger:(nonnull NSObject<FlutterBinaryMessenger> *)messenger {
@@ -11,7 +14,7 @@
     if (self) {
         
         _streamHandler = [[AppsFlyerStreamHandler alloc] init];
-        
+        _callbackChannel = [FlutterMethodChannel methodChannelWithName:afCallbacksMethodChannel binaryMessenger:messenger];
         _eventChannel = [FlutterEventChannel eventChannelWithName:afEventChannel binaryMessenger:messenger];
         [_eventChannel setStreamHandler:_streamHandler];
     }
@@ -19,12 +22,13 @@
 }
 
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-    
     id<FlutterBinaryMessenger> messenger = [registrar messenger];
     
     FlutterMethodChannel *channel = [FlutterMethodChannel methodChannelWithName:afMethodChannel binaryMessenger:messenger];
+    FlutterMethodChannel *callbackChannel = [FlutterMethodChannel methodChannelWithName:afCallbacksMethodChannel binaryMessenger:messenger];
     AppsflyerSdkPlugin *instance = [[AppsflyerSdkPlugin alloc] initWithMessenger:messenger];
     [registrar addMethodCallDelegate:instance channel:channel];
+    [registrar addMethodCallDelegate:instance channel:callbackChannel];
 }
 
 - (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
@@ -34,8 +38,8 @@
     }else if([@"getSDKVersion" isEqualToString:call.method]){
         [self getSDKVersion:result];
     }
-    else if([@"trackEvent" isEqualToString:call.method]){
-        [self trackEventWithCall:call result:result];
+    else if([@"logEvent" isEqualToString:call.method]){
+        [self logEventWithCall:call result:result];
     }else if([@"waitForCustomerUserId" isEqualToString:call.method]){
         [self waitForCustomerId:call result:result];
     }else if([@"setUserEmails" isEqualToString:call.method]){
@@ -49,8 +53,8 @@
         //
     }else if([@"enableLocationCollection" isEqualToString:call.method]){
         //
-    }else if([@"stopTracking" isEqualToString:call.method]){
-        [self stopTracking:call result:result];
+    }else if([@"stop" isEqualToString:call.method]){
+        [self stop:call result:result];
     }else if([@"setIsUpdate" isEqualToString:call.method]){
         //
     }else if([@"setCustomerUserId" isEqualToString:call.method]){
@@ -67,72 +71,159 @@
         [self setHost:call result:result];
     }else if([@"setAdditionalData" isEqualToString:call.method]){
         [self setAdditionalData:call result:result];
-    }else if([@"validateAndTrackInAppPurchase" isEqualToString:call.method]){
-        [self validateAndTrackInAppPurchase:call result:result];
+    }else if([@"validateAndLogInAppPurchase" isEqualToString:call.method]){
+        [self validateAndLogInAppPurchase:call result:result];
     }else if([@"getAppsFlyerUID" isEqualToString:call.method]){
         [self getAppsFlyerUID:result];
-    }else if([@"generateInviteLink" isEqualToString:call.method]){
-        [self generateInviteLink:call result:result];
     }else if([@"setSharingFilter" isEqualToString:call.method]){
         [self setSharingFilter:call result:result];
     }else if([@"setSharingFilterForAllPartners" isEqualToString:call.method]){
         [self setSharingFilterForAllPartners:result];
+    }else if([@"generateInviteLink" isEqualToString:call.method]){
+        [self generateInviteLink:call result:result];
+    }else if([@"setAppInviteOneLinkID" isEqualToString:call.method]){
+        [self setAppInviteOneLinkID:call result:result];
+    }else if([@"logCrossPromotionImpression" isEqualToString:call.method]){
+        [self logCrossPromotionImpression:call result:result];
+    }else if([@"logCrossPromotionAndOpenStore" isEqualToString:call.method]){
+        [self logCrossPromotionAndOpenStore:call result:result];
+    }else if([@"startListening" isEqualToString:call.method]){
+        [self startListening:call result:result];
+    }else if([@"setOneLinkCustomDomain" isEqualToString:call.method]){
+        [self setOneLinkCustomDomain:call result:result];
     }
     else{
         result(FlutterMethodNotImplemented);
     }
 }
 
+- (void)setOneLinkCustomDomain:(FlutterMethodCall*)call result:(FlutterResult)result{
+    NSArray* brandDomains = call.arguments;
+    [[AppsFlyerLib shared] setOneLinkCustomDomains:brandDomains];
+    result(nil);
+}
+
+- (void)startListening:(FlutterMethodCall*)call result:(FlutterResult)result{
+    // Prepare callback dictionary
+    if (self->callbackById == nil) self->callbackById = [NSMutableArray array];
+
+    NSString* callbackId = call.arguments;
+    [self->callbackById addObject:callbackId];
+}
+
 - (void)generateInviteLink:(FlutterMethodCall*)call result:(FlutterResult)result{
+    NSString* customerID = call.arguments[@"customerID"];
+    NSString* referrerImageUrl = call.arguments[@"referrerImageUrl"];
+    NSString* brandDomain = call.arguments[@"brandDomain"];
+    NSString* baseDeeplink = call.arguments[@"baseDeeplink"];
+    NSString* referrerName = call.arguments[@"referrerName"];
+    NSString* channel = call.arguments[@"channel"];
+    NSString* campaign = call.arguments[@"campaign"];
+    
     [AppsFlyerShareInviteHelper generateInviteUrlWithLinkGenerator:^AppsFlyerLinkGenerator * _Nonnull(AppsFlyerLinkGenerator * _Nonnull generator) {
-        [generator setChannel:call.arguments[@"channel"]];
+        [generator setChannel:channel];
+        [generator setCampaign:campaign];
+        [generator setBrandDomain:brandDomain];
+        [generator setBaseDeeplink:baseDeeplink];
+        [generator setReferrerName:referrerName];
+        [generator setReferrerImageURL:referrerImageUrl];
+        [generator setReferrerCustomerId:customerID];
+        
         return generator;
     } completionHandler:^(NSURL * _Nullable url) {
-        result([url absoluteString]);
+        NSString * resultURL = url.absoluteString;
+                    if(resultURL != nil){
+                        if([self->callbackById containsObject:@"successGenerateInviteLink"]){
+                        [self->_callbackChannel invokeMethod:@"callListener" arguments:@{
+                            @"id": @"successGenerateInviteLink",
+                            @"data":resultURL
+                        }];
+                        }
+                    }
     }];
+    
+    result(nil);
+}
+
+- (void)setAppInviteOneLinkID:(FlutterMethodCall*)call result:(FlutterResult)result{
+    NSString* oneLinkID = call.arguments[@"oneLinkID"];
+    [AppsFlyerLib shared].appInviteOneLinkID = oneLinkID;
+    result(nil);
+}
+
+- (void)logCrossPromotionImpression:(FlutterMethodCall*)call result:(FlutterResult)result{
+    NSString* appId = call.arguments[@"appId"];
+    NSString* campaign = call.arguments[@"campaign"];
+    NSDictionary* parameters = call.arguments[@"data"];
+    
+    [AppsFlyerCrossPromotionHelper logCrossPromoteImpression:appId campaign:campaign parameters:parameters];
+}
+
+- (void)logCrossPromotionAndOpenStore:(FlutterMethodCall*)call result:(FlutterResult)result{
+    NSString* campaign = call.arguments[@"campaign"];
+    NSDictionary* customParams = call.arguments[@"params"];
+    
+    [AppsFlyerShareInviteHelper generateInviteUrlWithLinkGenerator:^AppsFlyerLinkGenerator * _Nonnull(AppsFlyerLinkGenerator * _Nonnull generator) {
+                if (campaign != nil && ![campaign isEqualToString:@""]) {
+                    [generator setCampaign:campaign];
+                }
+                if (![customParams isKindOfClass:[NSNull class]]) {
+                    [generator addParameters:customParams];
+                }
+
+                return generator;
+            } completionHandler: ^(NSURL * _Nullable url) {
+                NSString *appLink = url.absoluteString;
+                if (@available(iOS 10.0, *)) {
+                    [[UIApplication sharedApplication] openURL:[NSURL URLWithString:appLink] options:@{} completionHandler:^(BOOL success) {
+                    }];
+                } else {
+                    // Fallback on earlier versions
+                }
+            }];
 }
 
 - (void)setSharingFilter:(FlutterMethodCall*)call result:(FlutterResult)result{
     NSArray* filters = call.arguments;
-    [[AppsFlyerTracker sharedTracker] setSharingFilter:filters];
+    [[AppsFlyerLib shared] setSharingFilter:filters];
     result(nil);
 }
 
 - (void)setSharingFilterForAllPartners:(FlutterResult)result{
-    [[AppsFlyerTracker sharedTracker] setSharingFilterForAllPartners];
+    [[AppsFlyerLib shared] setSharingFilterForAllPartners];
     result(nil);
 }
 
 - (void)getAppsFlyerUID:(FlutterResult)result{
-    result([[AppsFlyerTracker sharedTracker] getAppsFlyerUID]);
+    result([[AppsFlyerLib shared] getAppsFlyerUID]);
 }
 
 - (void)getHostPrefix:(FlutterResult)result{
-    result([[AppsFlyerTracker sharedTracker] hostPrefix]);
+    result([[AppsFlyerLib shared] hostPrefix]);
 }
 
 - (void)getHostName:(FlutterResult)result{
-    result([[AppsFlyerTracker sharedTracker] host]);
+    result([[AppsFlyerLib shared] host]);
 }
 
 - (void)getSDKVersion:(FlutterResult)result{
-    result([[AppsFlyerTracker sharedTracker] getSDKVersion]);
+    result([[AppsFlyerLib shared] getSDKVersion]);
 }
 
 - (void)setHost:(FlutterMethodCall*)call result:(FlutterResult)result{
     NSString* hostName = call.arguments[@"hostName"];
     NSString* hostPrefix = call.arguments[@"hostPrefix"];
-    [[AppsFlyerTracker sharedTracker] setHost:hostName withHostPrefix:hostPrefix];
+    [[AppsFlyerLib shared] setHost:hostName withHostPrefix:hostPrefix];
     result(nil);
 }
 
-- (void)validateAndTrackInAppPurchase:(FlutterMethodCall*)call result:(FlutterResult)result{
+- (void)validateAndLogInAppPurchase:(FlutterMethodCall*)call result:(FlutterResult)result{
     NSString* publicKey = call.arguments[@"publicKey"];
     NSString* signature = call.arguments[@"signature"];
     NSString* price = call.arguments[@"price"];
     NSString* currency = call.arguments[@"currency"];
     NSDictionary* additionalParameters = call.arguments[@"additionalParameters"];
-    [[AppsFlyerTracker sharedTracker] validateAndTrackInAppPurchase:publicKey price:price currency:currency transactionId:signature additionalParameters:additionalParameters
+    [[AppsFlyerLib shared] validateAndLogInAppPurchase:publicKey price:price currency:currency transactionId:signature additionalParameters:additionalParameters
                                                             success:^(NSDictionary *response) {
                                                                 NSLog(@"Success");
                                                                 [self onValidateSuccess:response];
@@ -165,19 +256,19 @@
 
 - (void)setAdditionalData:(FlutterMethodCall*)call result:(FlutterResult)result{
     NSDictionary* data = call.arguments[@"customData"];
-    [[AppsFlyerTracker sharedTracker] setAdditionalData:data];
+    [[AppsFlyerLib shared] setAdditionalData:data];
     result(nil);
 }
 
 - (void)setCustomerUserId:(FlutterMethodCall*)call result:(FlutterResult)result{
     NSString* userId = call.arguments[@"id"];
-    [[AppsFlyerTracker sharedTracker] setCustomerUserID:userId];
+    [[AppsFlyerLib shared] setCustomerUserID:userId];
     result(nil);
 }
 
-- (void)stopTracking:(FlutterMethodCall*)call result:(FlutterResult)result{
-    BOOL stopTracking = call.arguments[@"isTrackingStopped"];
-    [[AppsFlyerTracker sharedTracker] setIsStopTracking:stopTracking];
+- (void)stop:(FlutterMethodCall*)call result:(FlutterResult)result{
+    BOOL stop = call.arguments[@"isStopped"];
+    [AppsFlyerLib shared].isStopped = stop;
     result(nil);
 }
 
@@ -194,29 +285,36 @@
     NSArray *emaillsArray = [emails copy];
     NSNumber* cryptTypeInt = (id)call.arguments[@"cryptType"];
     EmailCryptType cryptType = (EmailCryptType)[cryptTypeInt integerValue];
-    [[AppsFlyerTracker sharedTracker] setUserEmails:emaillsArray withCryptType:cryptType];
+    [[AppsFlyerLib shared] setUserEmails:emaillsArray withCryptType:cryptType];
     result(nil);
 }
 
 - (void)setUserEmails:(FlutterMethodCall*)call result:(FlutterResult)result{
     NSMutableArray *emails = call.arguments[@"emails"];
     NSArray *emailsArray = [emails copy];
-    [[AppsFlyerTracker sharedTracker] setUserEmails:emailsArray withCryptType:EmailCryptTypeNone];
+    [[AppsFlyerLib shared] setUserEmails:emailsArray withCryptType:EmailCryptTypeNone];
     result(nil);
 }
 
 - (void)initSdkWithCall:(FlutterMethodCall*)call result:(FlutterResult)result{
-    
     NSString* devKey = nil;
     NSString* appId = nil;
+    NSString* appInviteOneLink = nil;
+    BOOL disableCollectASA = NO;
+    BOOL disableAdvertisingIdentifier = NO;
+    NSTimeInterval timeToWaitForATTUserAuthorization = 0;
     BOOL isDebug = NO;
     BOOL isConversionData = NO;
     
     id isDebugValue = nil;
     id isConversionDataValue = nil;
+    id isDisableCollectASA = nil;
+    id isDisableAdvertisingIdentifier = nil;
+
     devKey = call.arguments[afDevKey];
     appId = call.arguments[afAppId];
-    
+    timeToWaitForATTUserAuthorization = [(id)call.arguments[afTimeToWaitForATTUserAuthorization] doubleValue];
+
     isDebugValue = call.arguments[afIsDebug];
     if ([isDebugValue isKindOfClass:[NSNumber class]]) {
         // isDebug is a boolean that will come through as an NSNumber
@@ -228,38 +326,56 @@
     }
     
     if (isConversionData == YES) {
-        [[AppsFlyerTracker sharedTracker] setDelegate:_streamHandler];
+        [[AppsFlyerLib shared] setDelegate:_streamHandler];
     }
     
-    [AppsFlyerTracker sharedTracker].appleAppID = appId;
-    [AppsFlyerTracker sharedTracker].appsFlyerDevKey = devKey;
-    [AppsFlyerTracker sharedTracker].isDebug = isDebug;
+    appInviteOneLink = call.arguments[afInviteOneLink];
+    if(appInviteOneLink != nil){
+        [AppsFlyerLib shared].appInviteOneLinkID = appInviteOneLink;
+    }
+    
+    isDisableCollectASA = call.arguments[afDisableCollectASA];
+    if ([isDisableCollectASA isKindOfClass:[NSNumber class]]) {
+        // isDebug is a boolean that will come through as an NSNumber
+        disableCollectASA = [(NSNumber*)isDisableCollectASA boolValue];
+    }
+    isDisableAdvertisingIdentifier = call.arguments[afDisableAdvertisingIdentifier];
+    if ([isDisableAdvertisingIdentifier isKindOfClass:[NSNumber class]]) {
+        // isDebug is a boolean that will come through as an NSNumber
+        disableAdvertisingIdentifier = [(NSNumber*)isDisableAdvertisingIdentifier boolValue];
+    }
+    
+    [AppsFlyerLib shared].disableCollectASA = disableCollectASA;
+    [AppsFlyerLib shared].disableAdvertisingIdentifier = disableAdvertisingIdentifier;
+    [AppsFlyerLib shared].appleAppID = appId;
+    [AppsFlyerLib shared].appsFlyerDevKey = devKey;
+    [AppsFlyerLib shared].isDebug = isDebug;
     NSString *appInviteOneLinkID = call.arguments[afOneLinkId];
     [AppsFlyerTracker sharedTracker].appInviteOneLinkID = appInviteOneLinkID;
-    [[AppsFlyerTracker sharedTracker] trackAppLaunch];
+    [[AppsFlyerLib shared] start];
     
     [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(appDidBecomeActive) name:UIApplicationDidBecomeActiveNotification object:nil];
     
     result(@{@"status": @"OK"});
 }
 
--(void)trackEventWithCall:(FlutterMethodCall*)call result:(FlutterResult)result{
+-(void)logEventWithCall:(FlutterMethodCall*)call result:(FlutterResult)result{
     NSString *eventName =  call.arguments[afEventName];
     NSDictionary *eventValues = call.arguments[afEventValues];
     
-    [[AppsFlyerTracker sharedTracker] trackEvent:eventName withValues:eventValues];
+    [[AppsFlyerLib shared] logEvent:eventName withValues:eventValues];
     //TODO: Add callback handler
     result(@YES);
 }
 
 - (void)setMinTimeBetweenSessions:(FlutterMethodCall*)call result:(FlutterResult)result{
     NSInteger seconds = [(id)call.arguments[@"seconds"] integerValue];
-    [AppsFlyerTracker sharedTracker].minTimeBetweenSessions = seconds;
+    [AppsFlyerLib shared].minTimeBetweenSessions = seconds;
     result(nil);
 }
 
 - (void)appDidBecomeActive {
-    [[AppsFlyerTracker sharedTracker] trackAppLaunch];
+    [[AppsFlyerLib shared] start];
     NSLog(@"App Did Become Active");
 }
 
@@ -272,18 +388,18 @@
 
 -(void) handleCallback:(NSArray *) objArray{
     NSDictionary* message = [objArray objectAtIndex:0];
-    NSString* channel = [objArray objectAtIndex:1];
+    //NSString* channel = [objArray objectAtIndex:1];
     
     NSError *error;
     NSData *dataFromDict = [NSJSONSerialization dataWithJSONObject:message
                                                            options:NSJSONWritingPrettyPrinted
                                                              error:&error];
     [[NSNotificationCenter defaultCenter] postNotificationName:@"af-events" object:dataFromDict];
-//    if(!error){
-//        [flutterViewController sendOnChannel:channel message:dataFromDict binaryReply:^(NSData * _Nullable reply) {
-//            //
-//        }];
-//    }
+    //if(!error){
+        //[flutterViewController sendOnChannel:channel message:dataFromDict binaryReply:^(NSData * _Nullable reply) {
+            //
+        //}];
+    //}
 }