+++ lib/src/appsflyer_constants.dart
@@ -11,6 +11,8 @@ class AppsflyerConstants {
   static const String AF_DEV_KEY = "afDevKey";
   static const String AF_APP_Id = "afAppId";
   static const String AF_IS_DEBUG = "isDebug";
+  static const String AF_TIME_TO_WAIT_FOR_ATT_USER_AUTHORIZATION =
+      "timeToWaitForATTUserAuthorization";
   static const String AF_GCD = "GCD";
   static const String AF_SUCCESS = "success";
   static const String AF_FAILURE = "failure";
@@ -20,7 +22,11 @@ class AppsflyerConstants {
   static const String AF_EVENTS_CHANNEL = "af-events";
   static const String AF_METHOD_CHANNEL = "af-api";
 
-  static const AF_VALIDATE_PURCHASE = "validatePurchase";
+  static const String AF_VALIDATE_PURCHASE = "validatePurchase";
+  static const String APP_INVITE_ONE_LINK = "appInviteOneLink";
 
+  static const String DISABLE_COLLECT_ASA = "disableCollectASA";
+      "disableAdvertisingIdentifier";
+      
   static const String AF_ADDITIONAL_ONELINK_ID = "oneLinkId";
 }