+++ lib/src/appsflyer_options.dart
@@ -4,8 +4,18 @@ class AppsFlyerOptions {
   final String afDevKey;
   final bool showDebug;
   final String appId;
+  final double timeToWaitForATTUserAuthorization;
   final String appInviteOneLink;
+  final bool disableAdvertisingIdentifier;
+  final bool disableCollectASA;
 
-  AppsFlyerOptions(
-      {@required this.afDevKey, this.showDebug = false, this.appId = "", this.appInviteOneLink});
+  AppsFlyerOptions({
+    @required this.afDevKey,
+    this.showDebug = false,
+    this.appId = "",
+    this.timeToWaitForATTUserAuthorization,
+    this.appInviteOneLink,
+    this.disableAdvertisingIdentifier,
+    this.disableCollectASA,
+  });
 }